feat: update user imageUrl handling, add lastLogin tracking, and modify user migration for bigInt support

+++ src/migrations/20250611153622_create_users_and_refresh_tokens.ts
@@ -2,7 +2,7 @@ import { Knex } from 'knex';
 
 export async function up(knex: Knex): Promise<void> {
   await knex.schema.createTable('users', (table) => {
-    table.increments('id').primary();
+    table.bigIncrements('id').primary();
     table.string('name').notNullable();
     table.string('username').notNullable();
     table.string('email').notNullable().unique();
@@ -27,7 +27,7 @@ export async function up(knex: Knex): Promise<void> {
   await knex.schema.createTable('refresh_tokens', (table) => {
     table.increments('id').primary();
     table
-      .integer('user_id')
+      .bigInteger('user_id')
       .unsigned()
       .notNullable()
       .references('id')
